go an extra mile to find mged and set up paths properly regardless of the compile being in or out of dir -- untested and there are undoubtedly other commands, but it's one step
svn:revision:32206
svn:branch:trunk
svn:account:brlcad

+++ regress/iges.sh
@@ -1,14 +1,32 @@
 #!/bin/sh
 
-LD_LIBRARY_PATH=../src/other/tcl/unix:../src/other/tk/unix:$LD_LIBRARY_PATH
-DYLD_LIBRARY_PATH=../src/other/tcl/unix:../src/other/tk/unix:$DYLD_LIBRARY_PATH
+# Ensure /bin/sh
+export PATH || (echo "This isn't sh."; sh $0 $*; kill $$)
+
+# save the precious args
+ARGS="$*"
+NAME_OF_THIS=`basename $0`
+PATH_TO_THIS=`dirname $0`
+THIS="$PATH_TO_THIS/$NAME_OF_THIS"
+
+MGED="$1/src/mged/mged"
+if test ! -f "$MGED" ; then
+    MGED="$PATH_TO_THIS/../src/mged/mged"
+    if test ! -f "$MGED" ; then
+	echo "Unable to find mged, aborting"
+	exit 1
+    fi
+fi
+
+LD_LIBRARY_PATH=../src/other/tcl/unix:../src/other/tk/unix:$1/src/other/tcl/unix:$1/src/other/tk/unix:$LD_LIBRARY_PATH
+DYLD_LIBRARY_PATH=../src/other/tcl/unix:../src/other/tk/unix:$1/src/other/tcl/unix:$1/src/other/tk/unix:$DYLD_LIBRARY_PATH
 export LD_LIBRARY_PATH DYLD_LIBRARY_PATH
 
 rm -f iges.log iges.g iges_file.iges iges_stdout_new.g iges_new.g iges_stdout.iges iges_file.iges
 
 STATUS=0
 
-../src/mged/mged -c 2> iges.log <<EOF
+$MGED -c 2> iges.log <<EOF
 opendb iges.g y
 
 units mm
